% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{main_sim}
\alias{main_sim}
\title{Runs the main simulation}
\usage{
main_sim(n, items, times, parallel = TRUE, model_sim, ..., Sigma = NULL,
  weights = NULL)
}
\arguments{
\item{n,}{items: numeric vectors}

\item{times:}{numeric}

\item{model_sim:}{function to simulate Rasch data}

\item{...:}{additional arguments to model_sim}
}
\value{
returns a list that stores objects of class 'sim_res'
each array stores #times simulations of each possible n x items combination
}
\description{
Runs the main simulation
}
\examples{
main_sim(c(100, 400, 1000), c(10, 50, 70), sim.2pl, .4)
}

