% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{sim_removal}
\alias{sim_removal}
\title{Compute rejection of Rasch model based on Likelihood Ratio test after
items have been removed on the basis of item infit statistics}
\usage{
sim_removal(n, items, times, model_sim, ..., cores = detectCores(),
  cutoff = c(0.8, 1.2))
}
\arguments{
\item{n,}{items, times: numeric (single value)}

\item{model_sim:}{function to simulate Rasch data}

\item{...:}{additional arguments to model_sim}

\item{cores:}{number of cores to run in parallel}

\item{cutoff:}{item infit statistic cutoff}
}
\value{
returns an object of class 'sim_res'
which is an array that stores #times #items x 3 matrices
}
\description{
Compute rejection of Rasch model based on Likelihood Ratio test after
items have been removed on the basis of item infit statistics
}
\examples{
sim_removal(100, 20, sim.2pl, .4)
}

