% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{visualize}
\alias{visualize}
\title{Visualizes the % items removed given a simulation result}
\usage{
visualize(res_sum)
}
\arguments{
\item{res_sum:}{summarized simulation result}
}
\value{
plots a trellis graph visualising the results
}
\description{
Visualizes the % items removed given a simulation result
}
\examples{
spl <- summarize(sim_res)
visualize(spl)
}

